Fix book class and apply reviewer requirements
