+++ classes/book.rb
@@ -1,0 +1,19 @@
+require_relative '../item'
+
+class Book < Item
+  attr_accessor :publisher, :cover_state
+
+  def initialize(publisher, cover_state, published_date, archived)
+    super(published_date, archived)
+    @id = rand(1..1000)
+    @publisher = publisher
+    @cover_state = cover_state
+  end
+
+  def can_be_archived?
+    super || @cover_state == "bad"
+  end
+end
+
+book = Book.new("Publisher", "good", "2021-01-01", false)
+p book.can_be_archived?
